BotApp/requirements.txt
python-dotenv
aiohttp==3.9.5
asyncio==3.4.3
aiogram==3.3.0